% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usethis.R
\name{data}
\alias{data}
\title{Simulated data set for functionality tests testing}
\format{
A data frame with 1000 rows and 4 columns
}
\source{
Simulated
}
\description{
Some simulated data drawn from a diffuse state-space model
}
