% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FastOptim}
\alias{FastOptim}
\title{FastOptim Function}
\usage{
FastOptim(
  startR,
  XR,
  tauR,
  ind_matrix_R,
  day_indicator_R,
  rel_xtol,
  rel_ftol,
  max_eval,
  algorithm_id,
  compute_se,
  no_of_bootstraps,
  length_of_bootstraps,
  seed,
  hessian,
  step_size,
  log
)
}
\arguments{
\item{startR}{Starting values}

\item{XR}{Data Matrix}

\item{tauR}{Parameter vector}

\item{ind_matrix_R}{Identifier matrix}

\item{rel_xtol}{Algorithm tolerance}

\item{rel_ftol}{Stopping rule}

\item{algorithm_id}{NLopt algorithm ID}

\item{compute_se}{Whether or not standard eroors should be computed}

\item{no_of_bootstraps}{Number of bootstrap repititions}

\item{length_of_bootstraps}{Length of each bootstrap block}

\item{seed}{Seed used for bootstrapping}

\item{hessian}{Whether or not to compute the hessian}

\item{step_size}{Step size for the computation of the hessian}

\item{log}{Whether or not to print output}
}
\value{
Returns optimisation object
}
\description{
Fast KFS MLE implementation
}
