% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuoteDynamics.R
\docType{data}
\name{nlopt_algorithms}
\alias{nlopt_algorithms}
\title{Available Optimization Algorithms for FastOptim}
\format{
An object of class \code{numeric} of length 13.
}
\usage{
nlopt_algorithms
}
\description{
A named vector of available optimization algorithms for use with the `FastOptim` function.
}
\keyword{datasets}
