% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{objFunctionCpp}
\alias{objFunctionCpp}
\title{Wrapper to compute the neg-log-like ought to be called from within R for usage with Rs optim()}
\usage{
objFunctionCpp(startR, XR, tauR, ind_matrix_R, day_indicator_R, log)
}
\arguments{
\item{n}{Size of the paramter vector}

\item{x}{Pointer to the data of the parameter vector}

\item{grad}{Pointer to the data of the gradient}

\item{data}{Pointer to the data structure that holds additional function parameters of the actual objective function}
}
\value{
Objective function value of the current parameter estimate
}
\description{
Wrapper to compute the neg-log-like ought to be called from within R for usage with Rs optim()
}
