% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLoptKFS.R
\name{PrintAlgorithms}
\alias{PrintAlgorithms}
\title{Print Available NLopt Algorithms and Descriptions}
\usage{
PrintAlgorithms()
}
\description{
Lists the names of available optimization algorithms in `nlopt_algorithms`
along with a short description of each.
}
\examples{
PrintAlgorithms()
}
