% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\docType{package}
\name{QuoteDynamics}
\alias{QuoteDynamics}
\title{SPDX-License-Identifier: GPL-3.0-or-later */}
\description{
Copyright (c) 2025 Domenic Franjic
}
\details{
This file is part of QuoteDynamics.

QuoteDynamics is free software: you can redistribute it
and/or modify it under the terms of the GNU General Public License
as published by the Free Software Foundation, either version 3
of the License, or (at your option) any later version.

QuoteDynamics is distributed in the hope that it will be useful,
but WITHOUT ANY WARRANTY; without even the implied warranty
of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
See the GNU General Public License for more details.

You should have received a copy of the GNU General Public License
along with QuoteDynamics.  If not, see <https://www.gnu.org/licenses/>.
}
