% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuoteDynamics.R
\name{quoteDynNegLL}
\alias{quoteDynNegLL}
\title{Compute the negative log-likelihood of the quote dynamics model}
\usage{
quoteDynNegLL(start, X, tau, ident_mat, verbose = FALSE)
}
\arguments{
\item{start}{Vector of starting values.}

\item{X}{Matrix of data.}

\item{tau}{Vector of parameter values.}

\item{ident_mat}{Selection matrix.}

\item{verbose}{Whether to print verbose output.}
}
\value{
Negative log-likelihood
}
\description{
Compute the negative log-likelihood of the quote dynamics model
}
\examples{
start <- c(1, 1)
data <- matrix(rnorm(100), ncol = 10)
tau <- runif(10)
ident_mat <- matrix(round(runif(100 * 10, -0.5, 1.5), 0), 100, 10)
quoteDynNegLL(start, data, tau, ident_mat)
}
