% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuoteDynamics.R
\name{summary.quoteDynFit}
\alias{summary.quoteDynFit}
\title{Summary function for the return object of the optimisation call}
\usage{
\method{summary}{quoteDynFit}(object, ...)
}
\description{
Summary function for the return object of the optimisation call
}
