% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FastOptim}
\alias{FastOptim}
\title{FastOptim Function}
\usage{
FastOptim(
  startR,
  XR,
  tauR,
  ind_matrix_R,
  xtol,
  stop_val,
  max_eval,
  algorithm_id,
  hessian,
  step_size,
  log
)
}
\arguments{
\item{startR}{Starting values}

\item{XR}{Data Matrix}

\item{tauR}{Parameter vector}

\item{ind_matrix_R}{Identifier matrix}

\item{xtol}{Algorithm tolerance}

\item{stop_val}{Stopping rule}
}
\value{
Returns minimum value (changes start in place)
}
\description{
Fast KFS MLE implementation
}
